Split tests into multiple files to keep length reasonable
